requirements.txt
nump
simplejpeg
pyzmq
opencv-python
pyqt5
